requirements_ci.txt
attr==0.3.1
black==20.8b1
dm_tree==0.1.5
pandas==1.2.0
tensorboard==2.4.0
tensorflow_federated==0.17.0
tree==0.2.4
scikit-learn==0.24.0